fix: attention and fail-focus fixes fail-focus always on use ticks instead of milliseconds instead fix package naming to follow naming convention
