+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/aiming/Attention.kt
@@ -2,7 +2,9 @@ package net.ccbluex.liquidbounce.utils.aiming
 
 import net.ccbluex.liquidbounce.config.ToggleableConfigurable
 import net.ccbluex.liquidbounce.event.Listenable
-import net.ccbluex.liquidbounce.utils.client.Chronometer
+import net.ccbluex.liquidbounce.event.events.GameTickEvent
+import net.ccbluex.liquidbounce.event.handler
+import net.ccbluex.liquidbounce.utils.kotlin.EventPriorityConvention
 
 /**
  * The Attention modulates the rotation speed based on the time duration
@@ -15,12 +17,12 @@ class Attention(owner: Listenable? = null)
     : ToggleableConfigurable(owner, "Attention", false) {
 
     // Configuration properties
-    private val slowStartFactor by float("SlowStartFactor", 0.6f, 0.05f..0.7f)
-    private val transitionDuration by intRange("TransitionDuration", 250..500, 0..5000,
-        "ms")
+    private val slowStartFactor by float("SlowStartFactor", 0.6f, 0.01f..0.99f)
+    private val transitionDuration by intRange("TransitionDuration", 1..3, 1..20,
+        "ticks")
 
-    // A chronometer to track the duration for which the current target has been focused on
-    private val focusChronometer = Chronometer()
+    // A tick meter to track the duration for which the current target has been focused on
+    private var ticksElapsed = 0
 
     // The currently set transition duration, randomized within the defined range
     private var currentTransitionDuration = transitionDuration.random()
@@ -35,19 +37,24 @@ class Attention(owner: Listenable? = null)
                 return 1f
             }
 
-            val elapsed = focusChronometer.elapsed
-            return if (elapsed < currentTransitionDuration) {
+            val elapsed = ticksElapsed
+            return if (elapsed <= currentTransitionDuration) {
                 slowStartFactor + (1 - slowStartFactor) * (elapsed.toFloat() / currentTransitionDuration.toFloat())
             } else {
                 1f
             }
         }
 
+    @Suppress("unused")
+    private val gameHandler = handler<GameTickEvent>(priority = EventPriorityConvention.FIRST_PRIORITY) {
+        ticksElapsed++
+    }
+
     /**
      * Resets the chronometer and sets a new randomized transition duration when a new target is acquired.
      */
     fun onNewTarget() {
         currentTransitionDuration = transitionDuration.random()
-        focusChronometer.reset()
+        ticksElapsed = 0
     }
 }
